restruct, outgoing pkt len bug fix
