+++ lang/go/dz.ru/mqttudp/Parse.go
@@ -1,7 +1,6 @@
 package mqttudp
 
 import (
-	"fmt"
 	"net"
 	"reflect"
 )
@@ -39,18 +38,18 @@ Call callback function with resulting packet.
 	@param from_ip   Source IP address of packet.
 	@param acceptor  User function to call on packet parsed.
 */
-func Parse_any_pkt(raw []byte, from_ip *net.Addr, acceptor MqttUdpInput) error {
+func ParseBinary(raw []byte, from_ip *net.Addr) (MqttPacket, error) {
 	var plen = len(raw)
 	var err error = nil
 
 	var o MqttPacket
-	var pkt = 0
-
-	if plen < 2 {
-		return GlobalErrorHandler(Proto, "packet len < 2", "")
-	}
+	var pkt = 0 // packet parse position
 
 	o.Clear()
+
+	if plen < 2 {
+		return o, GlobalErrorHandler(Proto, "packet len < 2", "")
+	}
 
 	o.from_ip = from_ip
 	o.packetType = PType(raw[0])
@@ -62,7 +61,7 @@ func Parse_any_pkt(raw []byte, from_ip *net.Addr, acceptor MqttUdpInput) error {
 	o.value = nil
 
 	if o.total+2 > plen {
-		return GlobalErrorHandler(Proto, "packet too short", "")
+		return o, GlobalErrorHandler(Proto, "packet too short", "")
 	}
 
 	var ttrs_start int = pkt + o.total // end of payload, start of TTRs
@@ -86,11 +85,11 @@ func Parse_any_pkt(raw []byte, from_ip *net.Addr, acceptor MqttUdpInput) error {
 	pkt += 2
 
 	if tlen > MAX_SZ {
-		return GlobalErrorHandler(Proto, "packet too long", "")
+		return o, GlobalErrorHandler(Proto, "packet too long", "")
 	}
 
 	if pkt+tlen > o.total+2 {
-		return GlobalErrorHandler(Proto, "packet topic len > pkt len", "")
+		return o, GlobalErrorHandler(Proto, "packet topic len > pkt len", "")
 	}
 
 	o.topic = raw[pkt : pkt+tlen]
@@ -99,7 +98,7 @@ func Parse_any_pkt(raw []byte, from_ip *net.Addr, acceptor MqttUdpInput) error {
 
 	vlen = o.total - pkt + 2
 	if vlen > MAX_SZ {
-		return GlobalErrorHandler(Proto, "packet value len > pkt len", "")
+		return o, GlobalErrorHandler(Proto, "packet value len > pkt len", "")
 	}
 
 	// Packet with value?
@@ -110,23 +109,23 @@ func Parse_any_pkt(raw []byte, from_ip *net.Addr, acceptor MqttUdpInput) error {
 	o.value = raw[pkt : pkt+vlen]
 
 parse_ttrs:
-	;
+
 	var ttrs = raw[ttrs_start:] // Current position in TTRs
-	var ttrs_len = plen - ttrs_start
+	//var ttrs_len = plen - ttrs_start
 
-	//printf("TTRs  len=%d, plen=%d\n", ttrs_len, plen );
+	//fmt.Printf("TTRs  len=%d, plen=%d\n", len(ttrs), plen)
 
-	for ttrs_len > 0 {
+	for len(ttrs) > 0 {
 		var ttr_type = ttrs[0]
-		var ttr_pos = 1
+		var ttr_pos = 1 // skip type
 
 		var ttr_len = decode_size(ttrs, &ttr_pos)
 
 		if ttr_len <= 0 {
-			return GlobalErrorHandler(Proto, "TTR len < 0", "")
+			return o, GlobalErrorHandler(Proto, "TTR len <= 0", "")
 		}
 
-		//printf("TTR type = %c 0x%X len=%d\n", ttr_type, ttr_type, ttr_len );
+		//fmt.Printf("TTR type = %c 0x%X len=%d\n", ttr_type, ttr_type, ttr_len)
 		// Have TTR, process it
 		switch ttr_type {
 		case 'n':
@@ -136,10 +135,10 @@ parse_ttrs:
 			o.reply_to = ttr_decode_int32(ttrs[ttr_pos:])
 			break
 		case 's':
-			if ttr_len <= 0 {
-				err = GlobalErrorHandler(Proto, "signature TTR len != 16", "")
+			if ttr_len < 16 {
+				err = GlobalErrorHandler(Proto, "signature TTR len < 16", "")
 				if err != nil {
-					return err
+					return o, err
 				}
 			}
 			o.is_signed = false // TODO TTR_check_signature(pstart, ttr_start-pstart, ttrs)
@@ -149,24 +148,52 @@ parse_ttrs:
 			break
 		}
 
-		ttrs_len -= ttr_pos // type & len fields
-		ttrs_len -= ttr_len // TTR data
+		ttr_pos += ttr_len
 
-		ttrs = ttrs[ttr_len+ttr_pos:]
+		//ttrs_len -= ttr_pos // type & len fields
+		//ttrs_len -= ttr_len // TTR data
 
-		if ttrs_len < 0 {
-			return GlobalErrorHandler(Proto, "TTRs len < 0", "")
+		//ttrs = ttrs[ttr_len+ttr_pos:]
+
+		if len(ttrs) < ttr_pos+2 {
+			//return o, GlobalErrorHandler(Proto, "TTR len < 2", "")
+			// Ignore junk at end of pkt
+			return o, nil
 		}
+
+		ttrs = ttrs[ttr_pos:]
+
+		/*if ttrs_len < 0 {
+			return o, GlobalErrorHandler(Proto, "TTRs len < 0", "")
+		}*/
 	}
 
-	recv_reply(&o)
+	return o, err
+}
+
+/*
+Parse incoming packet.
+
+Call callback function with resulting packet.
+
+	@param raw       Incoming binary packet data from UDP packet.
+	@param from_ip   Source IP address of packet.
+	@param acceptor  User function to call on packet parsed.
+*/
+func ParseAndProcess(raw []byte, from_ip *net.Addr, acceptor MqttUdpInput) error {
+	o, err := ParseBinary(raw, from_ip)
+
+	if err != nil {
+		return err
+	}
+
 	o.call_packet_listeners()
 
 	if acceptor != nil {
 		acceptor.Accept(o)
 	}
 
-	return err
+	return nil
 }
 
 // -----------------------------------------------------------------------
@@ -220,70 +247,4 @@ func ttr_check_signature(buf []byte, pkt_len int, in_signature []byte) bool {
 	}
 
 	return true // TODO must return ok?
-}
-
-// -----------------------------------------------------------------------
-//
-// dump
-//
-// -----------------------------------------------------------------------
-
-var ptname = []string{"?0x00",
-	"CONNECT", "CONNACK", "PUBLISH", "PUBACK",
-	"PUBREC", "PUBREL", "PUBCOMP", "SUBSCRIBE",
-	"SUBACK", "UNSUBSCRIBE", "UNSUBACK", "PINGREQ",
-	"PINGRESP", "DISCONNECT",
-	"?0xF0"}
-
-// Dump packet.
-func (o *MqttPacket) Dump() {
-	var tn = ptname[o.packetType>>4]
-
-	var from = "(src unknown)"
-
-	if o.from_ip != nil {
-		from = (*o.from_ip).String()
-	}
-
-	fmt.Printf("pkt %10s flags %x, id %8x from %s",
-		tn, o.packetFlags, o.pkt_id, from)
-	/*int(0xFF&(o.from_ip>>24)),	int(0xFF&(o.from_ip>>16)),	int(0xFF&(o.from_ip>>8)),	int(0xFF&(o.from_ip))) */
-
-	if len(o.topic) > 0 {
-		fmt.Printf(" topic '%s'", o.topic)
-	}
-
-	if len(o.value) > 0 {
-		fmt.Printf(" = '%s'", o.value)
-	}
-
-	if o.is_signed {
-		fmt.Printf(" SIGNED")
-	}
-
-	fmt.Printf("\n")
-}
-
-/*
-Default packet processing, called from Parse_any_pkt()
-
-Reply to ping
-
-	@todo Reply to SUBSCRIBE? Not sure.
-	@todo Reply with PUBACK for PUBLISH with QoS
-	@todo Error handling
-*/
-func recv_reply(pkt *MqttPacket) {
-
-	switch pkt.packetType {
-	case PINGREQ:
-		// TODO err check
-		//if( fd > 0 ) mqtt_udp_send_ping_responce( fd, pkt->from_ip );
-		// TODO mqtt_udp_send_ping_responce()
-		break
-	//case PTYPE_SUBSCRIBE:
-	//case PTYPE_PUBLISH:
-	default:
-		break
-	}
 }
