+++ lang/go/dz.ru/mqttudp/parse_test.go
@@ -18,13 +18,13 @@ func (s myServer) Accept(packet MqttPacket) {
 func Test_Parse_Publish(t *testing.T) {
 	var raw = []byte{0x30, 8, 0, 4, 'a', 'b', 'c', 'd', 'x', 'y'}
 
-	Parse_any_pkt(raw, nil, nil)
+	ParseBinary(raw, nil)
 }
 
 func Test_Parse_Publish_TTR(t *testing.T) {
 	var raw = []byte{0x30, 8, 0, 4, 'a', 'b', 'c', 'd', 'x', 'y' /*TTR*/, 'n', 4, 0, 0, 0, 1}
 
-	Parse_any_pkt(raw, nil, nil)
+	ParseBinary(raw, nil)
 }
 
 func Test_Parse_Publish_callback(t *testing.T) {
@@ -33,7 +33,7 @@ func Test_Parse_Publish_callback(t *testing.T) {
 	var cb myServer
 	cb.t = t
 
-	Parse_any_pkt(raw, nil, cb)
+	ParseAndProcess(raw, nil, cb)
 }
 
 func Test_Parse_Publish_TTR_callback(t *testing.T) {
@@ -42,5 +42,5 @@ func Test_Parse_Publish_TTR_callback(t *testing.T) {
 	var cb myServer
 	cb.t = t
 
-	Parse_any_pkt(raw, nil, cb)
+	ParseAndProcess(raw, nil, cb)
 }
